refactor(gateway): extract User out of the lib
User (registry, routing, lifecycle, transport, flush) moves from the
de-mls lib into the gateway crate as the reference integrator. The lib
keeps the per-conversation SessionRunner and gains a small public
integrator API (process_inbound, attach_mls, has_mls) so User no longer
reaches into core internals. ConsensusContext stays in the lib (a
ConversationDeps dependency). User-driven tests move to gateway/tests.
